requirement.txt
polygon-pascalvoc-writer
xmltodict
opencv-python
tqdm
PyYAML
pillow